backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
pydantic==2.10.4
pydantic-settings==2.7.1
groq==0.15.0
httpx==0.28.1
pinecone-client==5.0.1
python-dotenv==1.0.1
tenacity==9.0.0
